requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
joblib==1.5.1
magic-filter==1.0.12
multidict==6.6.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.3.1
openai==1.99.1
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.19.2
pyTelegramBotAPI==4.28.0
pytest==8.4.1
python-dotenv==1.1.1
redis==6.4.0
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
textblob==0.19.0
tqdm==4.67.1
tweepy==4.16.0
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
tonsdk
setproctitle
psutil